break out pkg/scm/downloaders packages
